backend/requirements.txt
asgiref                       
bleach                       
certifi                       
cffi                          
charset-normalizer            
cryptography                  
defusedxml                    
Django                       
django-cors-headers          
django-filter                 
django-templated-mail         
djangorestframework          
djangorestframework-simplejwt 
djoser                        
idna                         
mysqlclient                  
oauthlib                      
pillow                       
pip                         
pycparser                     
PyJWT                        
python-dotenv                 
python3-openid                
pytz                         
requests                     
requests-oauthlib             
six                           
social-auth-app-django        
social-auth-core             
sqlparse                    
tzdata                      
urllib3                      
webencodings              